+++ geotrek/feedback/static/feedback/main.js
@@ -1,3 +1,35 @@
+//
+// Feedback / reports
+//
+
+$(window).on('entity:map', function (e, data) {
+    var modelname = 'report';
+    var layername = `${modelname}_layer`;
+    var url = window.SETTINGS.urls[layername];
+    var loaded_infrastructure = false;
+    var map = data.map;
+
+    // Show report layer in application maps
+    var layer = new L.ObjectsLayer(null, {
+        modelname: modelname,
+        style: L.Util.extend(window.SETTINGS.map.styles[modelname] || {}, { clickable: false }),
+    });
+
+    if (data.modelname != modelname) {
+        map.layerscontrol.addOverlay(layer, tr('Report'), tr('Feedback'));
+    };
+
+    map.on('layeradd', function (e) {
+        var options = e.layer.options || { 'modelname': 'None' };
+        if (!loaded_infrastructure) {
+            if (options.modelname == modelname && options.modelname != data.modelname) {
+                e.layer.load(url);
+                loaded_infrastructure = true;
+            }
+        }
+    });
+});
+
 $(window).on('entity:view:add entity:view:update', function (e, data) {
     $('#div_id_message_sentinel').prop('hidden', true);
     $('#div_id_message_supervisor').prop('hidden', true);